Added webpack to the project and reorganized the project structure.
Added a package.json, .gitignore, webpack.config.js and yarn.lock file to set up webpack.
Moved all the files into the public folder and created a src and dist folder in the js folder to handle builds.
Removed index.html and renamed startscreen.html to index.html.
Changed the path for the script file in the new index.html.